Refactor: Move TableViewUtil.findScrollbar into BisqTableView
